+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Changelog
-
 ## v0.4.0-alpha
-- Initial Build 1 Part A
+- Kernel
+- Container
+- EventBus
+- Memory Engine